# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GAPSI API",
    "version": "1.0.0",
    "description": "API para gestionar el catálogo de GAPSI."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/articles": {
      "post": {
        "summary": "Crear nuevo artículo",
        "operationId": "createArticle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nombre del artículo (hasta 20 caracteres).",
                    "example": "Artículo de prueba"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descripción del artículo (hasta 200 caracteres).",
                    "example": "Esta es una descripción de ejemplo del artículo."
                  },
                  "price": {
                    "type": "number",
                    "format": "float",
                    "description": "Precio del artículo (dos decimales).",
                    "example": 100.5
                  },
                  "model": {
                    "type": "string",
                    "description": "Identificador del modelo (hasta 10 caracteres).",
                    "example": "MODELO1234"
                  }
                },
                "required": ["name", "price"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artículo creado con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta debido a errores de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "get": {
        "summary": "Obtener todos los artículos",
        "operationId": "getAllArticles",
        "responses": {
          "200": {
            "description": "Una lista de artículos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "summary": "Obtener un artículo por ID",
        "operationId": "getArticleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del artículo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artículo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "404": {
            "description": "Artículo no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Artículo no encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar un artículo",
        "operationId": "updateArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del artículo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "Descripción actualizada del artículo (hasta 200 caracteres).",
                    "example": "Descripción actualizada."
                  },
                  "model": {
                    "type": "string",
                    "description": "Identificador del modelo actualizado (hasta 10 caracteres).",
                    "example": "ACTUALIZADO1234"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Artículo actualizado con éxito",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            },
            "404": {
              "description": "Artículo no encontrado",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Artículo no encontrado"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Solicitud incorrecta debido a errores de validación",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        },
        "delete": {
          "summary": "Eliminar un artículo",
          "operationId": "deleteArticle",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID del artículo",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Artículo eliminado con éxito"
            },
            "404": {
              "description": "Artículo no encontrado",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Artículo no encontrado"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador único del artículo (10 caracteres).",
            "example": "ABCDEFGHIJ"
          },
          "name": {
            "type": "string",
            "description": "Nombre del artículo (hasta 20 caracteres).",
            "example": "Artículo de muestra"
          },
          "description": {
            "type": "string",
            "description": "Descripción del artículo (hasta 200 caracteres).",
            "example": "Esta es una descripción de ejemplo del artículo."
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Precio del artículo (dos decimales).",
            "example": 100.5
          },
          "model": {
            "type": "string",
            "description": "Identificador del modelo (hasta 10 caracteres).",
            "example": "MODELO1234"
          }
        }
      }
    }
  }
}